more config changes
git-svn-id: http://svn.xp-dev.com/svn/wxextension@1893 f22100f3-aa73-48fe-a4fb-fd497bb32605

+++ trunk/syncedit/src/frame.cpp
@@ -619,11 +619,11 @@ void MDIFrame::OnCommand(wxCommandEvent& event)
     break;
 
   case ID_OPTION_LIST_SORT_ASCENDING:
-    wxExApp::SetConfig("List/SortMethod", SORT_ASCENDING); break;
+    wxExApp::SetConfig("List/SortMethod", (long)SORT_ASCENDING); break;
   case ID_OPTION_LIST_SORT_DESCENDING:
-    wxExApp::SetConfig("List/SortMethod", SORT_DESCENDING); break;
+    wxExApp::SetConfig("List/SortMethod", (long)SORT_DESCENDING); break;
   case ID_OPTION_LIST_SORT_TOGGLE:
-    wxExApp::SetConfig("List/SortMethod", SORT_TOGGLE); break;
+    wxExApp::SetConfig("List/SortMethod", (long)SORT_TOGGLE); break;
 
   case ID_PROCESS_SELECT: wxExProcessWithListView::ConfigDialog(this); break;
